+++ src/main/java/com/ecotrack/carbon_tracker/service/UserService.java
@@ -1,6 +1,7 @@
 package com.ecotrack.carbon_tracker.service;
 
 import com.ecotrack.carbon_tracker.entity.User;
+import com.ecotrack.carbon_tracker.exception.ResourceNotFoundException;
 import com.ecotrack.carbon_tracker.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -24,29 +25,33 @@ public class UserService {
         return userRepository.findByUsername(username);
     }
 
-    public List<User> getAllUser () {
+    public List<User> getAllUser() {
         return userRepository.findAll();
     }
 
-    public User getUserById (Long id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User Not Found!"));
-
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
     }
 
     public User updateUser(Long id, User newUser) {
-        User user = userRepository.findById(id).get();
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
+
         user.setUsername(newUser.getUsername());
         user.setRole(newUser.getRole());
 
-        if(user.getPassword() != null && !user.getPassword().isEmpty()) {
+        if (newUser.getPassword() != null && !newUser.getPassword().isEmpty()) {
             user.setPassword(passwordEncoder.encode(newUser.getPassword()));
         }
 
         return userRepository.save(user);
     }
 
-    public void deleteUser (long id) {
+    public void deleteUser(long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
